// code/frontend/apps/game-index/src/gen/catalogApi/types/MostAnticipatedGame.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type MostAnticipatedGame = {
    /**
     * @type integer, int64
    */
    id: number;
    /**
     * @type string
    */
    name: string;
    /**
     * @type null,string
    */
    coverUrl?: string | null;
    /**
     * @type string
    */
    gameType: string;
    /**
     * @type integer, int32
    */
    hypes: number;
    /**
     * @type string, date-time
    */
    firstReleaseDateUtc: string;
};